pyproject.toml
[tool.poetry]
name = "wews"
version = "0.1.0"
description = ""
authors = ["Andrés Garrido <andreslp001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^0.19.9"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.104.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
